change from dpkg-deb to dpkg-buildpackage
add debian directory and move the respective files there and add
missing (rules, compat).

Move control.in to control as no sed search & replacement must be
done. Add a Source section to the control file.

Rename the "check" Makefile target to "test" so that the dh_auto_test
helper executes it automatically on build.
Move the verify_api check to this target.

Cleanup the "clean" target and remove outdated commented out stuff

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>